Refactor AddressViews to be shared utilities
