+++ server/models/view-models/utils.ts
@@ -1,4 +1,5 @@
 import { ErrorSummary } from '../../utils/govukFrontEndTypes/errorSummary'
+import { Address, AddressTypeEnum } from '../Address'
 
 export type ErrorMessage = {
   text: string
@@ -10,18 +11,6 @@ export type FormField = {
 
 export type TextField = FormField & {
   value: string
-}
-
-export type Address = {
-  line1: string
-  line2: string
-  line3: string
-  line4: string
-  postcode: string
-}
-
-export type AddressField = FormField & {
-  value: Address
 }
 
 export type Time = {
@@ -70,6 +59,30 @@ export type ViewModel<T> = {
   [K in keyof T]: T[K] extends Date ? DateField : T[K] extends string[] ? MultipleChoiceField : TextField
 }
 
+export type AddressViews = {
+  primaryAddressView: string
+  secondaryAddressView: string
+  tertiaryAddressView: string
+}
+export type AddressViewsViewModel<T> = ViewModel<T> & {
+  primaryAddressView: TextField
+  secondaryAddressView: TextField
+  tertiaryAddressView: TextField
+}
+
+export const getAddressViews = (addresses: Address[]): AddressViews => {
+  const primaryAddress = addresses?.find(address => address.addressType === AddressTypeEnum.Enum.PRIMARY)
+  const secondaryAddress = addresses?.find(address => address.addressType === AddressTypeEnum.Enum.SECONDARY)
+  const tertiaryAddress = addresses?.find(address => address.addressType === AddressTypeEnum.Enum.TERTIARY)
+
+  const addressViews = {
+    primaryAddressView: primaryAddress ? primaryAddress.addressLine1 : '',
+    secondaryAddressView: secondaryAddress ? secondaryAddress.addressLine1 : '',
+    tertiaryAddressView: tertiaryAddress ? tertiaryAddress.addressLine1 : '',
+  }
+
+  return addressViews
+}
 export type ErrorsViewModel = {
   [field: string]: ErrorMessage
 }
